for `images`
It doesnt really seem to work. Im not sure if its the code or the sql syntax that I dont know about. I just couldnt insert images into the database directly. Not with load_files(), not with hex() or anything that I found online. maybe I will change stuff up to see if it works.